Updated module augmentation for ckbox package.
