+++ packages/ckeditor5-ckbox/src/ckboxcommand.ts
@@ -425,9 +425,3 @@ type CKBoxEvent<Name extends '' | 'choose' | 'open' | 'close' = ''> = {
 	name: Name extends '' ? 'ckbox' : `ckbox:${ Name }`;
 	args: Name extends 'choose' ? [ assets: Array<CKBoxRawAssetDefinition> ] : [];
 };
-
-declare module '@ckeditor/ckeditor5-core' {
-	interface CommandsMap {
-		ckbox: CKBoxCommand;
-	}
-}
